[Property wrapers] Fix l-value computation in buildStorageReference
Prior to this commit, buildStorageReference used to take an 'isLValue' argument
that specified whether the reference should be an l-value or not.

This provided sufficient information in case of a stored property
(self.member). However, when accessing a wrapped property
(self._member.wrappedValue), this doesn't have enough information to decide
whether intermediate components of the expression should be l-values or not.
Prior to this commit, this was extrapolated from the 'isLValue' argument,
leading to compiler crashes or unnecessary intermediate l-values when using
non-standard mutability with property wrappers.

To fix this, this commit replaces the 'isLValue' argument with two arguments:

  isUsedForGetAccess: Specifies that the referenced storage is meant to be
                      accessed with a getter
  isUsedForSetAccess: Specifies that the referenced storage is meant to be
                      accessed with a setter

and updates buildStorageReference to correctly figure out l-value-ness of
self and intermediate components from these arguments.