Force load all members of iterable contexts in the DWARF Importer
The DWARF Importer does not yet know how to answer a request at the level of a single member.  All the tests that saw this as an observable behavior were actually seeing a complete cache fill from lazy member loading an initializer fail over to flushing the member table and rebuilding it by forcing all the members. Given that the cache flush is something we're trying to avoid in general, this is obviously undesirable behavior.

Now that we're no longer flushing the cache for initializers, the DWARF Importer needs to simulate the old behavior and completely deserialize members of loaded types.  This pessimization is justified anyways, considering if you're loading from DWARF, you're probably about to print the type and its fields anyways.