[Experiment] Stick a cache in front of lazy member loading
Add a cache of lazily-imported names so we don't run off and deserialize extensions multiple times.  The cache indicates that the lookup table has a complete understanding of any given base name.  As a consequence, it must be flushed when a new extension with lazy members is added to avoid returning inconsistent results.

This should make lazy member cache misses much, much cheaper. In the best case, we'll avoid repeatedly crawling around on disk.  In the average case, we'll have fallen off the lazy member loading path at some point for some extension and the lazily-complete cache will kick in to keep that one extension from pessimizing the rest.  In the worst case - when an enormous amount of lookups for non-existent members occur - we'll probably balloon memory usage somewhat adding bogus members to the set.