Sema: Don't mark property wrapper 'modify' accessors transparent
There are certain conditions that cause them to reference
non-public properties, in particular if the property
wrapper is implemented via a subscript taking a keypath.

Instead of trying to detect this specific case though (and
possibly missing others), I'm just going to decree that
'modify' is never going to be transparent for property
wrappers.

Fixes <rdar://problem/57609867>.