[semantic-arc-opts] When performing load [copy] -> load_borrow on classes, do not ignore forwarding uses.
This is the first of two commits. This commit is a very simple, easily
cherry-pickable fix but does not use the LiveRange infrastructure so that we
handle forwarding uses here. Instead, we just bail if all uses of our load
[copy] are not destroy_value.

In a subsequent commit, I am going to change this to use the LiveRange
infrastructure so we will handle these cases. Sadly doing so doesn't cherry-pick
well. = /.

rdar://58289320
(cherry picked from commit 23f36a059a15f47b2a622bf16cb89557a97aa18c)