[sil] Add a new builder for Arguments called SILArgumentBuilder and use that in SILBasicBlock to create arguments instead of doing it manually.
With time, we are going to move to a world where SILBasicBlock is not in the
business of constructing SILArguments. With that in mind, this commit refactors
the impl in SILBasicBlock to use this builder instead. Naturally, this means
that SILBasicBlock is no longer a friend class of SILArgument, eliminating an
abstraction leak!

That being said, there are a ton of call sites that call these
SILBasicBlock::*Argument methods that we will need to fix over time and until
that happens, we can not eliminate the old method. This commit at least allows
us to stauch the bleeding and ensures that I can put create methods for some of
the new arguments I am adding, only onto the arg builder class.

Should be NFC.