[Experiment] Redo Globals As Member Tables
Globals-as-members has an antagonistic relationship with lazy member
loading. The old PCM extension filed them away indexed by context, which
was super efficient for the general case of "just import this entire
extension", but ridiculously inefficient for the specific case of "just
load this one member for me".

Instead, store the table in named order.  Switch over all the callers
and unblock lazy member loading.

This naturally changes the order things are printed in, but does not
change their contents.