[di] Teach DI how to mark uninitialized ref_element_addr with an uninitialized access bit.
This will enable us to assume that all ref_element_addr that have not been
marked as uninitialized are guaranteed to not be used by swift's DI for the
purposes of initializing uninitialized memory.

The reason why I am messing with this is that I want to enable safe interior
pointer verification in SIL and so I need to know whether or not a
ref_element_addr is guaranteed to point at something that is uninitialized or
not. Given that DI provides that guarantee for Swift classes, it seems
reasonable that this will let us encode this property.