Use LLDB.framework install dir provided by cmake cache
Allow the lldb build to find the correct install dir for LLDB.framework
in a cmake cache. (The path currently hardcoded into build-script is not
correct for embedded builds.)

To do this, build-script needs to tell lldb where to install each
arch-specific LLDB.framework, so it can lipo all these together for a
fat build. build-script does already do this with CMAKE_INSTALL_PREFIX,
but LLDB.framework is special, because it doesn't always share an
install prefix with other lldb products.

Depends on: https://reviews.llvm.org/D72286